+++ activemq-cpp/src/main/decaf/internal/nio/ShortArrayBuffer.cpp
@@ -31,7 +31,7 @@ ShortArrayBuffer::ShortArrayBuffer( int size, bool readOnly )
 
     // Allocate using the ByteArray, not read-only initially.  Take a reference to it.
     // The capacity is the given capacity times the size of the stored datatype
-    this->_array.reset( new ByteArrayAdapter( size * sizeof(short) ) );
+    this->_array.reset( new ByteArrayAdapter( size * (int)sizeof(short) ) );
     this->offset = 0;
     this->length = size;
     this->readOnly = readOnly;
