+++ activemq-cpp/src/test/activemq/util/MarshallingSupportTest.cpp
@@ -193,7 +193,7 @@ void MarshallingSupportTest::testWriteString() {
 
         MarshallingSupport::writeString( writer, "" );
 
-        bais.setByteArray( baos.toByteArray(), baos.size() );
+        bais.setByteArray( baos.toByteArray(), (int)baos.size() );
 
         CPPUNIT_ASSERT( dataIn.read() == PrimitiveValueNode::STRING_TYPE );
         CPPUNIT_ASSERT( dataIn.readShort() == 0 );
@@ -206,7 +206,7 @@ void MarshallingSupportTest::testWriteString() {
 
         MarshallingSupport::writeString( writer, "Hello World" );
 
-        bais.setByteArray( baos.toByteArray(), baos.size() );
+        bais.setByteArray( baos.toByteArray(), (int)baos.size() );
 
         CPPUNIT_ASSERT( dataIn.read() == PrimitiveValueNode::STRING_TYPE );
         CPPUNIT_ASSERT( dataIn.readShort() == 11 );
@@ -219,7 +219,7 @@ void MarshallingSupportTest::testWriteString() {
 
         MarshallingSupport::writeString( writer, std::string( Short::MAX_VALUE, 'A' ) );
 
-        bais.setByteArray( baos.toByteArray(), baos.size() );
+        bais.setByteArray( baos.toByteArray(), (int)baos.size() );
 
         CPPUNIT_ASSERT( dataIn.read() == PrimitiveValueNode::BIG_STRING_TYPE );
         CPPUNIT_ASSERT( dataIn.readInt() == Short::MAX_VALUE );
@@ -236,7 +236,7 @@ void MarshallingSupportTest::testWriteString16() {
         DataInputStream dataIn( &bais );
 
         MarshallingSupport::writeString16( writer, "Hello World" );
-        bais.setByteArray( baos.toByteArray(), baos.size() );
+        bais.setByteArray( baos.toByteArray(), (int)baos.size() );
 
         CPPUNIT_ASSERT( dataIn.readShort() == 11 );
     }
@@ -252,7 +252,7 @@ void MarshallingSupportTest::testWriteString32() {
         DataInputStream dataIn( &bais );
 
         MarshallingSupport::writeString32( writer, "Hello World" );
-        bais.setByteArray( baos.toByteArray(), baos.size() );
+        bais.setByteArray( baos.toByteArray(), (int)baos.size() );
 
         CPPUNIT_ASSERT( dataIn.readInt() == 11 );
     }
@@ -272,7 +272,7 @@ void MarshallingSupportTest::testReadString16() {
     MarshallingSupport::writeString( dataOut, testStr );
 
     // Move the output back to the input.
-    bytesIn.setByteArray( bytesOut.toByteArray(), bytesOut.size() );
+    bytesIn.setByteArray( bytesOut.toByteArray(), (int)bytesOut.size() );
 
     string resultStr = "";
     int type = dataIn.read();
@@ -300,7 +300,7 @@ void MarshallingSupportTest::testReadString32() {
     MarshallingSupport::writeString( dataOut, testStr );
 
     // Move the output back to the input.
-    bytesIn.setByteArray( bytesOut.toByteArray(), bytesOut.size() );
+    bytesIn.setByteArray( bytesOut.toByteArray(), (int)bytesOut.size() );
 
     string resultStr = "";
     int type = dataIn.read();
